post_flow/grouper-nfe/groupers/grouper-medicamentos/requirements.txt
requests
pandas
matplotlib
langchain-core
langchain-ollama
SPARQLWrapper 
pysolr 
pyyaml
SQLAlchemy
psycopg2-binary
sentry-sdk
pika
pyarrow 
boto3 
python-dotenv 
s3fs
pandas
Minio
